pyproject.toml
[project]
name = "rag-arena"
version = "0.1.0"
description = "RAGBench-12x: Reproducible benchmark for 12 RAG configurations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beir>=2.0.0",
    "bm25s>=0.2.14",
    "chromadb>=1.3.7",
    "langdetect>=1.0.9",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "stop-words>=2025.11.4",
    "tiktoken>=0.12.0",
    "langgraph>=0.1.0",
    "ragas>=0.2.0",
    "sentence-transformers>=3.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "langchain>=1.2.0",
    "langfuse>=3.11.2",
    "langchain-openai>=1.1.6",
    "cohere>=5.20.1",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]

[project.scripts]
ragbench = "ragbench.cli:main"

[tool.uv]
package = true
